API Base de Datos CRUD/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
mysql-connector-python==9.3.0
python-dotenv==1.1.1
PyJWT==2.10.1